+++ src/constants/editor.ts
@@ -1,5 +1,7 @@
 import { CodeEditorSuggestionItem, CodeEditorSuggestionItemKind } from '@grafana/ui';
 
+import { afterRenderCodeParameters, beforeRenderCodeParameters } from '../helpers';
+
 /**
  * Supported Languages
  */
@@ -34,86 +36,6 @@ export enum Format {
 export const FORMAT_OPTIONS = [
   { value: Format.AUTO, label: 'Auto', icon: 'font' },
   { value: Format.NONE, label: 'Disabled', icon: 'times-circle' },
-];
-
-/**
- * EDITOR CONTEXT SUGGESTIONS
- */
-export const EDITOR_CONTEXT_SUGGESTIONS: CodeEditorSuggestionItem[] = [
-  {
-    label: 'context',
-    kind: CodeEditorSuggestionItemKind.Constant,
-    detail: 'All passed possible properties and methods.',
-  },
-  {
-    label: 'context.data',
-    kind: CodeEditorSuggestionItemKind.Property,
-    detail: 'Row data.',
-  },
-  {
-    label: 'context.dataFrame',
-    kind: CodeEditorSuggestionItemKind.Property,
-    detail: 'Selected data frame.',
-  },
-  {
-    label: 'context.panelData',
-    kind: CodeEditorSuggestionItemKind.Property,
-    detail: 'Panel data.',
-  },
-
-  /**
-   * Grafana
-   */
-  {
-    label: 'context.grafana',
-    kind: CodeEditorSuggestionItemKind.Property,
-    detail: 'Grafana properties and methods.',
-  },
-  {
-    label: 'context.grafana.eventBus',
-    kind: CodeEditorSuggestionItemKind.Property,
-    detail: 'Panels events.',
-  },
-  {
-    label: 'context.grafana.locationService',
-    kind: CodeEditorSuggestionItemKind.Property,
-    detail: 'Location service.',
-  },
-  {
-    label: 'context.grafana.replaceVariables',
-    kind: CodeEditorSuggestionItemKind.Method,
-    detail: 'Interpolate variables.',
-  },
-  {
-    label: 'context.grafana.theme',
-    kind: CodeEditorSuggestionItemKind.Property,
-    detail: 'Grafana theme.',
-  },
-  {
-    label: 'context.grafana.notifySuccess',
-    kind: CodeEditorSuggestionItemKind.Method,
-    detail: 'Display successful notification.',
-  },
-  {
-    label: 'context.grafana.notifyError',
-    kind: CodeEditorSuggestionItemKind.Method,
-    detail: 'Display error notification.',
-  },
-  {
-    label: 'context.grafana.getLocale',
-    kind: CodeEditorSuggestionItemKind.Method,
-    detail: 'Returns locale.',
-  },
-  {
-    label: 'context.grafana.timeZone',
-    kind: CodeEditorSuggestionItemKind.Property,
-    detail: 'Selected time zone.',
-  },
-  {
-    label: 'context.grafana.timeRange',
-    kind: CodeEditorSuggestionItemKind.Property,
-    detail: 'Selected time range.',
-  },
 ];
 
 /**
@@ -172,35 +94,15 @@ export const HELPERS_EDITOR_SUGGESTIONS: CodeEditorSuggestionItem[] = [
   },
 
   /**
-   * Common Context Suggestions
+   * Context Suggestions
    */
-  ...EDITOR_CONTEXT_SUGGESTIONS,
-
-  /**
-   * Specific Context Suggestions
-   */
-  {
-    label: 'context.handlebars',
-    kind: CodeEditorSuggestionItemKind.Property,
-    detail: 'Handlebars library.',
-  },
+  ...beforeRenderCodeParameters.suggestions,
 ];
 
 /**
  * After Render Editor Suggestions
  */
-export const AFTER_RENDER_EDITOR_SUGGESTIONS: CodeEditorSuggestionItem[] = [
-  ...EDITOR_CONTEXT_SUGGESTIONS,
-
-  /**
-   * Specific Context Suggestions
-   */
-  {
-    label: 'context.element',
-    kind: CodeEditorSuggestionItemKind.Property,
-    detail: 'Row html element.',
-  },
-];
+export const AFTER_RENDER_EDITOR_SUGGESTIONS: CodeEditorSuggestionItem[] = afterRenderCodeParameters.suggestions;
 
 /**
  * Editor Type