+++ CHANGELOG.md
@@ -4,9 +4,11 @@
 
 ### Features / Enhancements
 
-- Update context (#270)
+- Update context parameter (#270)
 - Added theme object, notifySuccess & notifyError (#270)
 - Update dependencies and Actions (#271)
+- Replace custom code parameters with Code Parameters Builder (#285)
+- Update CSS class for the Panel instead of a Row (#272)
 
 ## 4.3.0 (2023-12-25)